Create acceptance test OccUsersGroupsContext
