+++ tests/acceptance/config/behat.yml
@@ -134,6 +134,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - EmailContext:
         - OccContext:
+        - OccUsersGroupsContext:
 
     cliMain:
       paths:
@@ -242,6 +243,7 @@ default:
         - WebUIPersonalSecuritySettingsContext:
         - WebUIUserContext:
         - OccContext:
+        - OccUsersGroupsContext:
 
     webUIRenameFiles:
       paths:
