+++ tests/acceptance/features/bootstrap/OccContext.php
@@ -199,64 +199,6 @@ class OccContext implements Context {
 	}
 
 	/**
-	 * @When /^the administrator creates (?:these users|this user) using the occ command:$/
-	 * @Given /^(?:these users have|this user has) been created using the occ command:$/
-	 * expects a table of users with the heading
-	 * "|username|password|displayname|email|"
-	 * displayname & email are optional
-	 *
-	 * @param TableNode $table
-	 *
-	 * @return void
-	 * @throws \Exception
-	 */
-	public function theseUsersHaveBeenCreatedUsingTheOccCommand(TableNode $table) {
-		foreach ($table as $row) {
-			$username = $row['username'];
-			$cmd = "user:add $username  --password-from-env";
-
-			if (isset($row['displayname'])) {
-				$displayName = $row['displayname'];
-			} else {
-				$displayName = $this->featureContext->getDisplayNameForUser($username);
-			}
-
-			if ($displayName !== null) {
-				$cmd = "$cmd --display-name='$displayName'";
-			}
-
-			if (isset($row['email'])) {
-				$email = $row['email'];
-			} else {
-				$email = $this->featureContext->getEmailAddressForUser($username);
-			}
-
-			if ($email !== null) {
-				$cmd = "$cmd --email='$email'";
-			}
-
-			if (isset($row['password'])) {
-				$password = $row['password'];
-			} else {
-				$password = $this->featureContext->getPasswordForUser($row ['username']);
-			}
-
-			$this->invokingTheCommandWithEnvVariable(
-				$cmd,
-				'OC_PASS',
-				$password
-			);
-
-			$this->featureContext->addUserToCreatedUsersList(
-				$username,
-				$password,
-				$displayName,
-				$email
-			);
-		}
-	}
-
-	/**
 	 * @Given the administrator has set the mail smtpmode to :smtpmode
 	 *
 	 * @param string $smtpmode
@@ -266,290 +208,6 @@ class OccContext implements Context {
 	public function theAdministratorHasSetTheMailSmtpmodeTo($smtpmode) {
 		$this->invokingTheCommand(
 			"config:system:set  --value $smtpmode mail_smtpmode"
-		);
-	}
-
-	/**
-	 * @When the administrator tries to create a user :username using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorTriesToCreateAUserUsingTheOccCommand($username) {
-		$this->invokingTheCommandWithEnvVariable(
-			"user:add $username  --password-from-env",
-			'OC_PASS',
-			$this->featureContext->getPasswordForUser($username)
-		);
-	}
-
-	/**
-	 * @When the administrator creates user :username password :password group :group using the occ command
-	 *
-	 * @param string $username
-	 * @param string $password
-	 * @param string $group
-	 *
-	 * @return void
-	 */
-	public function theAdministratorCreatesUserPasswordGroupUsingTheOccCommand($username, $password, $group) {
-		$cmd = "user:add $username  --password-from-env --group=$group";
-		$this->invokingTheCommandWithEnvVariable(
-			$cmd,
-			'OC_PASS',
-			$this->featureContext->getActualPassword($password)
-		);
-		$this->featureContext->addUserToCreatedUsersList($username, $password);
-	}
-
-	/**
-	 * @When the administrator resets the password of user :username to :password using the occ command
-	 *
-	 * @param string $username
-	 * @param string $password
-	 *
-	 * @return void
-	 */
-	public function resetUserPasswordUsingTheOccCommand($username, $password) {
-		$this->featureContext->resetUserPassword($username, $password);
-	}
-
-	/**
-	 * @When the administrator resets their own password to :newPassword using the occ command
-	 *
-	 * @param string $newPassword
-	 *
-	 * @return void
-	 */
-	public function theAdministratorResetsTheirOwnPasswordToUsingTheOccCommand($newPassword) {
-		$password = $this->featureContext->getActualPassword($newPassword);
-		$admin = $this->featureContext->getAdminUsername();
-		$this->invokingTheCommandWithEnvVariable(
-			"user:resetpassword $admin --password-from-env",
-			'OC_PASS',
-			$password
-		);
-		$this->featureContext->rememberNewAdminPassword($password);
-	}
-
-	/**
-	 * @When the administrator invokes password reset for user :username using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorInvokesPasswordResetForUserUsingTheOccCommand($username) {
-		$this->featureContext->resetUserPassword($username);
-	}
-
-	/**
-	 * @When the administrator changes the email of user :username to :newEmail using the occ command
-	 *
-	 * @param string $username
-	 * @param string $newEmail
-	 *
-	 * @return void
-	 */
-	public function theAdministratorChangesTheEmailOfUserToUsingTheOccCommand($username, $newEmail) {
-		$this->invokingTheCommand(
-			"user:modify $username email $newEmail"
-		);
-	}
-
-	/**
-	 * @When the administrator changes the display name of user :username to :newDisplayname using the occ command
-	 *
-	 * @param string $username
-	 * @param string $newDisplayname
-	 *
-	 * @return void
-	 */
-	public function theAdministratorChangesTheDisplayNameOfUserToUsingTheOccCommand($username, $newDisplayname) {
-		$this->invokingTheCommand(
-			"user:modify $username displayname '$newDisplayname'"
-		);
-	}
-
-	/**
-	 * @When the administrator changes the quota of user :username to :newQuota using the occ command
-	 *
-	 * @param string $username
-	 * @param string $newQuota
-	 *
-	 * @return void
-	 */
-	public function theAdministratorChangesTheQuotaOfUserToUsingTheOccCommand($username, $newQuota) {
-		$this->invokingTheCommand(
-			"user:modify $username quota $newQuota"
-		);
-	}
-
-	/**
-	 * @When the administrator deletes user :username using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorDeletesUserUsingTheOccCommand($username) {
-		$this->invokingTheCommand(
-			"user:delete $username"
-		);
-	}
-
-	/**
-	 * @When the administrator retrieves all the users in JSON format using the occ command
-	 *
-	 * @return void
-	 */
-	public function theAdministratorRetrievesAllTheUsersInJsonUsingTheOccCommand() {
-		$this->invokingTheCommand(
-			"user:list --output=json"
-		);
-	}
-
-	/**
-	 * @When the administrator retrieves the information of user :username in JSON format using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorRetrievesTheInformationOfUserInJsonUsingTheOccCommand($username) {
-		$this->invokingTheCommand(
-			"user:list $username --output=json"
-		);
-	}
-
-	/**
-	 * @When the administrator gets the groups of user :username in JSON format using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorGetsTheGroupsOfUserInJsonUsingTheOccCommand($username) {
-		$this->invokingTheCommand(
-			"user:list-group $username --output=json"
-		);
-	}
-
-	/**
-	 * @When the administrator retrieves the time when user :username was last seen using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorRetrievesTheTimeWhenUserWasLastSeenUsingTheOccCommand($username) {
-		$this->invokingTheCommand(
-			"user:lastseen $username"
-		);
-	}
-
-	/**
-	 * @When the administrator changes the language of user :username to :language using the occ command
-	 *
-	 * @param string $username
-	 * @param string $language
-	 *
-	 * @return void
-	 */
-	public function theAdministratorChangesTheLanguageOfUserToUsingTheOccCommand(
-		$username, $language
-	) {
-		$this->invokingTheCommand(
-			"user:setting $username core lang --value='$language'"
-		);
-	}
-
-	/**
-	 * @When the administrator retrieves the user report using the occ command
-	 *
-	 * @return void
-	 */
-	public function theAdministratorRetrievesTheUserReportUsingTheOccCommand() {
-		$this->invokingTheCommand("user:report");
-	}
-
-	/**
-	 * @When the administrator creates group :group using the occ command
-	 *
-	 * @param string $group
-	 *
-	 * @return void
-	 */
-	public function theAdministratorCreatesGroupUsingTheOccCommand($group) {
-		$this->invokingTheCommand(
-			"group:add $group"
-		);
-		$this->featureContext->addGroupToCreatedGroupsList($group);
-	}
-
-	/**
-	 * @When the administrator adds user :username to group :group using the occ command
-	 *
-	 * @param string $username
-	 * @param string $group
-	 *
-	 * @return void
-	 */
-	public function theAdministratorAddsUserToGroupUsingTheOccCommand($username, $group) {
-		$this->invokingTheCommand(
-			"group:add-member -m $username $group"
-		);
-	}
-
-	/**
-	 * @When the administrator deletes group :group using the occ command
-	 *
-	 * @param string $group
-	 *
-	 * @return void
-	 */
-	public function theAdministratorDeletesGroupUsingTheOccCommand($group) {
-		$this->invokingTheCommand(
-			"group:delete $group"
-		);
-	}
-
-	/**
-	 * @When the administrator gets the users in group :groupName in JSON format using the occ command
-	 *
-	 * @param string $groupName
-	 *
-	 * @return void
-	 */
-	public function theAdministratorGetsTheUsersInGroupInJsonUsingTheOccCommand($groupName) {
-		$this->invokingTheCommand(
-			"group:list-members $groupName --output=json"
-		);
-	}
-
-	/**
-	 * @When the administrator gets the groups in JSON format using the occ command
-	 *
-	 * @return void
-	 */
-	public function theAdministratorGetsTheGroupsInJsonUsingTheOccCommand() {
-		$this->invokingTheCommand(
-			"group:list --output=json"
-		);
-	}
-
-	/**
-	 * @When the administrator removes user :username from group :group using the occ command
-	 *
-	 * @param string $username
-	 * @param string $group
-	 *
-	 * @return void
-	 */
-	public function theAdministratorRemovesUserFromGroupUsingTheOccCommand($username, $group) {
-		$this->invokingTheCommand(
-			"group:remove-member -m $username $group"
 		);
 	}
 
@@ -614,48 +272,6 @@ class OccContext implements Context {
 		$this->invokingTheCommand(
 			"app:getpath $appName"
 		);
-	}
-
-	/**
-	 * @When the administrator disables user :username using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function theAdministratorDisablesUserUsingTheOccCommand($username) {
-		$this->invokingTheCommand(
-			"user:disable $username"
-		);
-	}
-
-	/**
-	 * @When administrator enables user :username using the occ command
-	 *
-	 * @param string $username
-	 *
-	 * @return void
-	 */
-	public function administratorEnablesUserUsingTheOccCommand($username) {
-		$this->invokingTheCommand(
-			"user:enable $username"
-		);
-	}
-
-	/**
-	 * @Then the language of user :username returned by the occ command should be :language
-	 *
-	 * @param string $username
-	 * @param string $language
-	 *
-	 * @return void
-	 */
-	public function theLanguageOfUserReturnedByTheOccCommandShouldBe($username, $language) {
-		$this->invokingTheCommand(
-			"user:setting $username core lang"
-		);
-		$responseLanguage = $this->featureContext->getStdOutOfOccCommand();
-		PHPUnit_Framework_Assert::assertEquals($language, \trim($responseLanguage));
 	}
 
 	/**
@@ -965,105 +581,6 @@ class OccContext implements Context {
 		foreach ($appsSimplified as $app) {
 			PHPUnit_Framework_Assert::assertContains($app, $lastOutputApps);
 		}
-	}
-
-	/**
-	 * @Then the users returned by the occ command should be
-	 *
-	 * @param TableNode $useridTable
-	 *
-	 * @return void
-	 */
-	public function theUsersReturnedByTheOccCommandShouldBe(TableNode $useridTable) {
-		$lastOutput = $this->featureContext->getStdOutOfOccCommand();
-		$lastOutputUsers = \json_decode($lastOutput, true);
-		$result = [];
-		// check if an array is a multi-dimensional array with inner array key 'displayName'
-		if (\array_column($lastOutputUsers, 'displayName')) {
-			foreach ($lastOutputUsers as $key => $value) {
-				$result[$key] =  $value['displayName'];
-			}
-		} else {
-			$result = $lastOutputUsers;
-		}
-		foreach ($useridTable as $row) {
-			PHPUnit_Framework_Assert::assertArrayHasKey($row['uid'], $result);
-			PHPUnit_Framework_Assert::assertContains($row['display name'], $result);
-		}
-	}
-
-	/**
-	 * @Then the groups returned by the occ command should be
-	 *
-	 * @param TableNode $groupTableNode with group name with header group
-	 *
-	 * @return void
-	 */
-	public function theGroupsReturnedByTheOccCommandShouldBe(TableNode $groupTableNode) {
-		$lastOutput = $this->featureContext->getStdOutOfOccCommand();
-		$lastOutputGroups = \json_decode($lastOutput, true);
-
-		foreach ($groupTableNode as $row) {
-			PHPUnit_Framework_Assert::assertContains($row['group'], $lastOutputGroups);
-		}
-	}
-
-	/**
-	 * @Then the display name returned by the occ command should be :displayName
-	 *
-	 * @param string $displayName
-	 *
-	 * @return void
-	 */
-	public function theDisplayNameReturnedByTheOccCommandShouldBe($displayName) {
-		$lastOutput = $this->featureContext->getStdOutOfOccCommand();
-		$lastOutputUser = \json_decode($lastOutput, true);
-		$lastOutputDisplayName = \array_column($lastOutputUser, 'displayName')[0];
-		PHPUnit_Framework_Assert::assertEquals($displayName, $lastOutputDisplayName);
-	}
-
-	/**
-	 * @Then the command output of user last seen should be recently
-	 *
-	 * @return void
-	 * @throws \Exception
-	 */
-	public function theCommandOutputOfUserLastSeenShouldBeRecently() {
-		$currentTime = \gmdate('d.m.Y H:i');
-		$currentTimeStamp = \strtotime($currentTime);
-		$lastOutput = $this->featureContext->getStdOutOfOccCommand();
-		\preg_match("/([\d.]+ [\d:]+)/", $lastOutput, $userCreatedTime);
-		$useCreatedTimeStamp = \strtotime(($userCreatedTime[0]));
-		$delta = $currentTimeStamp - $useCreatedTimeStamp;
-		if ($delta > 60) {
-			throw new Exception(__METHOD__ . "User was expected to be seen recently but wasn't");
-		}
-	}
-
-	/**
-	 * @Then the command output of user last seen should be never
-	 *
-	 * @return void
-	 */
-	public function theCommandOutputOfUserLastSeenShouldBeNever() {
-		$lastOutput = $this->featureContext->getStdOutOfOccCommand();
-		PHPUnit_Framework_Assert::assertContains(
-			"has never logged in.",
-			$lastOutput
-		);
-	}
-
-	/**
-	 * @Then the total users returned by the command should be :noOfUsers
-	 *
-	 * @param integer $noOfUsers
-	 *
-	 * @return void
-	 */
-	public function theTotalUsersReturnedByTheCommandShouldBe($noOfUsers) {
-		$lastOutput = $this->featureContext->getStdOutOfOccCommand();
-		\preg_match("/\|\s+total users\s+\|\s+(\d+)\s+\|/", $lastOutput, $actualUsers);
-		PHPUnit_Framework_Assert::assertEquals($noOfUsers, $actualUsers[1]);
 	}
 
 	/**
